# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Feriados do Brasil",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/estados": {
      "get": {
        "summary": "States",
        "operationId": "states_v1_estados_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/cobertura": {
      "get": {
        "summary": "Coverage",
        "operationId": "coverage_v1_cobertura_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/feriados": {
      "get": {
        "summary": "List Holidays",
        "operationId": "list_holidays_v1_feriados_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Ano"
            }
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaFeriados"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feriados/verificar": {
      "get": {
        "summary": "Check Holiday",
        "operationId": "check_holiday_v1_feriados_verificar_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "data",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Data"
            }
          },
          {
            "name": "uf",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 2,
              "title": "Uf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Feriado": {
        "properties": {
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "nacional",
              "estadual"
            ],
            "title": "Tipo"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "fonte": {
            "type": "string",
            "title": "Fonte"
          },
          "versao_fonte": {
            "type": "string",
            "title": "Versao Fonte"
          },
          "atualizado_em": {
            "type": "string",
            "title": "Atualizado Em"
          }
        },
        "type": "object",
        "required": [
          "data",
          "nome",
          "tipo",
          "uf",
          "fonte",
          "versao_fonte",
          "atualizado_em"
        ],
        "title": "Feriado"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListaFeriados": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "feriados": {
            "items": {
              "$ref": "#/components/schemas/Feriado"
            },
            "type": "array",
            "title": "Feriados"
          }
        },
        "type": "object",
        "required": [
          "ano",
          "uf",
          "total",
          "feriados"
        ],
        "title": "ListaFeriados"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}